friend

+++ src/main/java/com/harian/share/location/closersharelocation/user/UserDTO.java
@@ -1,5 +1,8 @@
 package com.harian.share.location.closersharelocation.user;
 
+import com.harian.share.location.closersharelocation.user.model.Gender;
+import com.harian.share.location.closersharelocation.user.model.User;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserDTO {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private String avatar;
